requirements.txt
# This version is per 21 Sept 2022
google-cloud-storage==2.5.0